requirements.txt
absl-py==1.4.0
actionlib==1.14.0
addict==2.4.0
angles==1.9.13
argcomplete==3.4.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
backcall==0.2.0
bagpy==0.5
base_local_planner==1.17.3
bitarray==3.3.1
bitstring==4.3.1
blinker==1.8.2
bondpy==1.8.6
boto3==1.35.91
botocore==1.35.91
cachetools==5.3.3
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cryptography==44.0.0
cv-bridge==1.16.2
cycler==0.12.1
Cython==3.0.12
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.7
decorator==4.4.2
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dynamic-reconfigure==1.7.3
einops==0.8.0
evo==1.31.1
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.8.0.post1
faiss-gpu==1.7.2
fast_pytorch_kmeans==0.2.2
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.12.23
fonttools==4.53.1
fsspec==2024.12.0
g2o-python==0.0.12
gast==0.3.3
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gdown==5.2.0
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gnupg==2.3.1
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
h5py==3.11.0
huggingface-hub==0.27.0
idna==3.10
image-geometry==1.16.2
imageio==2.34.2
importlib_metadata==8.0.0
importlib_resources==6.4.0
iniconfig==2.1.0
inputs==0.5
interactive-markers==1.12.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.4.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==2.13.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.4.7
kornia==0.7.3
kornia-moons==0.2.9
kornia_rs==0.1.8
laser_geometry==1.6.7
lazy_loader==0.4
libclang==18.1.1
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
MegEngine==1.9999.0+cu118
megfile==2.0.5
message-filters==1.16.0
mpmath==1.3.0
mprop==0.17.0
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
numexpr==2.8.6
numpy==1.24.4
numpy-quaternion==2023.0.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
open3d==0.18.0
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pillow_heif==0.18.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==5.5.1
pure_eval==0.2.3
py3rosmsgs==1.18.2
pyarrow==14.0.2
pybind11==2.13.1
pycparser==2.22
pydbow==0.1.2
pyflann-py3==0.1.0
pygame==2.6.0
pyglet==2.1.3
Pygments==2.18.0
PyOpenGL==3.1.7
pyOpenSSL==24.3.0
pyparsing==3.1.4
PyQt5-Qt5==5.15.16
PyQt5_sip==12.15.0
pyqtgraph==0.13.3
pyquaternion==0.9.9
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.8.2
python-qt-binding==0.4.4
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.1
pyzed @ file:///tmp/selfgz3411950/pyzed-4.2-cp38-cp38-linux_x86_64.whl#sha256=aeed3fe3dd4258a4e2d1aece6099ae0991441503d839b661883b3112effa8c8d
pyzmq==26.2.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rerun-sdk==0.21.0
resource_retriever==1.12.7
retrying==1.3.4
ros_numpy==0.0.5
rosbag==1.16.0
rosbags==0.9.23
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosserial_python==0.9.2
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rpds-py==0.20.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.9
rtabmap-python==0.21.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
rviz==1.14.25
s3transfer==0.10.4
safetensors==0.5.0
scikit-build==0.18.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sensor-msgs==1.13.1
setuptools-scm==8.2.0
six==1.17.0
sklearn-pandas==1.8.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tf==1.13.2
tf-conversions==1.13.2
tf-keras==2.15.0
tf-slim==1.1.0
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
tf2-sensor-msgs==0.7.7
threadpoolctl==3.5.0
tifffile==2023.7.10
tomli==2.0.1
topic-tools==1.16.0
torch==2.2.0
torchvision==0.17.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
triton==2.2.0
turtlebot3_teleop==1.2.5
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urdfdom-py==0.4.6
urllib3==2.2.3
wcwidth==0.2.13
werkzeug==2.2.2
widgetsnbextension==4.0.13
xacro==1.14.18
yacs==0.1.8
zipp==3.20.2
zstandard==0.23.0
